Rename renderers to views to better represent what they do.
