+++ src/client/systems/syscommand.cpp
@@ -20,7 +20,7 @@
 
 #include <string>
 
-#include "client/systems/sysstarsystemrenderer.h"
+#include "client/systems/views/starsystemview.h"
 #include "client/systems/gui/systooltips.h"
 #include "client/scenes/universescene.h"
 #include "common/components/bodies.h"
