+++ src/client/systems/sysstarsystemtree.cpp
@@ -18,7 +18,7 @@
 
 #include <string>
 
-#include "client/systems/sysstarsystemrenderer.h"
+#include "client/systems/views/starsystemview.h"
 #include "client/systems/gui/systooltips.h"
 #include "client/scenes/universescene.h"
 #include "common/components/bodies.h"
