+++ src/client/systems/views/galaxyview.h
